requirements.txt
python-docx>=1.1.0,<2.0.0
python-pptx>=1.0.0,<2.0.0
